+++ src/main/java/com/kaya/dto/response/FitnessReportResponse.java
@@ -12,9 +12,9 @@ import java.util.Set;
 @AllArgsConstructor
 public class FitnessReportResponse {
     private Long id;
-    private Long roomConflicts;
-    private Long instructorConflicts;
-    private Long studentConflicts;
-    private Long totalPenalty;
+    private Integer roomConflicts;
+    private Integer instructorConflicts;
+    private Integer studentConflicts;
+    private Integer totalPenalty;
     private Set<Lecture> conflictingLectures;
 }
